+++ opencivicdata/core/models/__init__.py
@@ -1,0 +1,10 @@
+# flake8: NOQA
+from .jurisdiction import Jurisdiction
+from .division import Division
+from .people_orgs import (
+    Organization, OrganizationIdentifier, OrganizationName, OrganizationContactDetail,
+    OrganizationLink, OrganizationSource,
+    Person, PersonIdentifier, PersonName, PersonContactDetail, PersonLink, PersonSource,
+    Post, PostContactDetail, PostLink,
+    Membership, MembershipContactDetail, MembershipLink
+)
