+++ opencivicdata/legislative/models/__init__.py
@@ -1,13 +1,5 @@
 # flake8: NOQA
-from .jurisdiction import Jurisdiction, LegislativeSession
-from .division import Division
-from .people_orgs import (
-    Organization, OrganizationIdentifier, OrganizationName, OrganizationContactDetail,
-    OrganizationLink, OrganizationSource,
-    Person, PersonIdentifier, PersonName, PersonContactDetail, PersonLink, PersonSource,
-    Post, PostContactDetail, PostLink,
-    Membership, MembershipContactDetail, MembershipLink
-)
+from .session import LegislativeSession
 from .bill import (Bill, BillAbstract, BillTitle, BillIdentifier, RelatedBill, BillSponsorship,
                    BillDocument, BillVersion, BillDocumentLink, BillVersionLink, BillSource,
                    BillActionRelatedEntity, BillAction)
