+++ opencivicdata/legislative/models/event.py
@@ -2,8 +2,10 @@ from __future__ import unicode_literals
 from django.contrib.gis.db import models
 from django.contrib.postgres.fields import ArrayField
 from django.utils.encoding import python_2_unicode_compatible
-from .base import OCDBase, LinkBase, OCDIDField, RelatedBase, MimetypeLinkBase, RelatedEntityBase
-from .jurisdiction import Jurisdiction
+from opencivicdata.core.models.base import (OCDBase, LinkBase, OCDIDField,
+                                            RelatedBase, RelatedEntityBase,
+                                            MimetypeLinkBase)
+from opencivicdata.core.models import Jurisdiction
 from .bill import Bill
 from .vote import VoteEvent
