+++ opencivicdata/legislative/models/vote.py
@@ -3,9 +3,9 @@ from django.db import models
 from django.contrib.postgres.fields import ArrayField
 from django.utils.encoding import python_2_unicode_compatible
 
-from .base import OCDBase, LinkBase, OCDIDField, RelatedBase
-from .people_orgs import Organization, Person
-from .jurisdiction import LegislativeSession
+from opencivicdata.core.models.base import OCDBase, LinkBase, OCDIDField, RelatedBase
+from opencivicdata.core.models import Organization, Person
+from .session import LegislativeSession
 from .bill import Bill, BillAction
 from .. import common
